feat: ServiceWriter added
