+++ src/main/java/utils/contentCreators/ServiceSaver.java
@@ -16,7 +16,7 @@ import java.security.NoSuchAlgorithmException;
 
 public class ServiceSaver {
     private final PassphraseHandler passphraseHandler = new PassphraseHandler();
-
+    private final ServiceWriter serviceWriter = new ServiceWriter();
 
     public ServiceData saveService(
             String serviceName,
@@ -48,43 +48,22 @@ public class ServiceSaver {
         if (generator == null) password = passwordReader.readLine("Enter password: ");
         else password = generator.generatePassword();
 
-        var serviceFile = getServiceFile(serviceName, contentFolder);
-        var serviceWriter = new FileWriter(serviceFile);
-        serviceWriter.write(encryptor.encrypt(login, passphrase) + "\n" + encryptor.encrypt(password, passphrase) + "\n");
+        var serviceData = new ServiceData(login, password);
+
+        serviceWriter.writeService(
+                encryptor.encrypt(login, passphrase) + "\n" + encryptor.encrypt(password, passphrase) + "\n",
+                serviceName,
+                contentFolder);
 
         System.out.println(encryptor.encrypt(login, passphrase) + "\n" + encryptor.encrypt(password, passphrase));
 
         serviceWriter.close();
 
-        return new ServiceData(login, password);
-    }
-
-    private File getServiceFile(String serviceName, String contentFolder) throws IOException {
-        var serviceFile = new File(contentFolder + File.separator + serviceName);
-
-        if (!serviceFile.getParentFile().exists()) if (!serviceFile.getParentFile().mkdirs())
-            throw new RuntimeException("Could not create folder " + serviceFile.getParentFile().getAbsolutePath());
-
-
-        if (serviceFile.exists()) {
-            throw new RuntimeException("Service " + serviceName + " already exists");
-//            if (new AgreementHandler().yesNoQuestion("Service already exists: " + serviceFile.getName() + "Want to rewrite password? (y/n)")) {
-//                System.out.println("TODO");
-//                // TODO
-//                //  ???? pls kirill v next time pishi chto nado todo
-//            }
-//
-//            return;
-        }
-
-        if (!serviceFile.createNewFile())
-            throw new RuntimeException("Could not create file " + serviceFile.getAbsolutePath());
-        return serviceFile;
+        return serviceData;
     }
 
     public boolean serviceExists(String serviceName, String contentFolder) {
-        var serviceFile = new File(contentFolder + File.separator + serviceName);
-        return serviceFile.exists();
+        return new File(contentFolder + File.separator + serviceName).exists();
     }
 
     public void updateService(String serviceName, String contentFolder) {
